Fix to #10293 - Invalid cast when using inline "if" returning an enum
Problem was for byte values in general - SQL server would return them as int and the client projection would expect byte, which lead to exception. Fix is to introduce SqlServer specific type mapping for byte that would produce a correct literal: CAST (value as tinyint). Same thing goes for short (int16) and long (int64).
This also reverts fix to 8538, which we get for free with the int64 change.